:recycle: api: migrate add certification candidate route to src
