--- api/lib/infrastructure/mail-check.js
@@ -1,2 +1,0 @@
-// FIXME this file should be deleted after migration
-export * from '../../src/shared/mail/infrastructure/services/mail-check.js';
