--- api/lib/infrastructure/utils/string-utils.js
@@ -1,2 +1,0 @@
-// FIXME this file should be deleted after migration
-export * from '../../../src/shared/infrastructure/utils/string-utils.js';
