+++ api/tests/certification/session/unit/domain/usecases/add-certification-candidate-to-session_test.js
@@ -1,13 +1,13 @@
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
-import { addCertificationCandidateToSession } from '../../../../lib/domain/usecases/add-certification-candidate-to-session.js';
-import { CpfBirthInformationValidation } from '../../../../lib/domain/services/certification-cpf-service.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
+import { addCertificationCandidateToSession } from '../../../../../../src/certification/session/domain/usecases/add-certification-candidate-to-session.js';
+import { CpfBirthInformationValidation } from '../../../../../../src/certification/shared/domain/services/certification-cpf-service.js';
 
 import {
   CertificationCandidateByPersonalInfoTooManyMatchesError,
   CertificationCandidateOnFinalizedSessionError,
   CertificationCandidatesError,
-} from '../../../../lib/domain/errors.js';
-import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../lib/domain/constants/certification-candidates-errors.js';
+} from '../../../../../../lib/domain/errors.js';
+import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../../lib/domain/constants/certification-candidates-errors.js';
 
 describe('Unit | UseCase | add-certification-candidate-to-session', function () {
   let certificationCandidateRepository;
