+++ api/tests/integration/domain/services/certification-candidates-ods-service/certification-candidates-ods-service_test.js
@@ -6,12 +6,12 @@ import {
   PIX_PLUS_EDU_2ND_DEGRE,
 } from '../../../../../lib/domain/models/ComplementaryCertification.js';
 import * as certificationCandidatesOdsService from '../../../../../lib/domain/services/certification-candidates-ods-service.js';
-import * as certificationCpfService from '../../../../../lib/domain/services/certification-cpf-service.js';
-import * as certificationCpfCountryRepository from '../../../../../lib/infrastructure/repositories/certification-cpf-country-repository.js';
-import * as certificationCpfCityRepository from '../../../../../lib/infrastructure/repositories/certification-cpf-city-repository.js';
+import * as certificationCpfService from '../../../../../src/certification/shared/domain/services/certification-cpf-service.js';
+import * as certificationCpfCountryRepository from '../../../../../src/certification/shared/infrastructure/repositories/certification-cpf-country-repository.js';
+import * as certificationCpfCityRepository from '../../../../../src/certification/shared/infrastructure/repositories/certification-cpf-city-repository.js';
 import * as certificationCenterRepository from '../../../../../src/certification/shared/infrastructure/repositories/certification-center-repository.js';
 import * as complementaryCertificationRepository from '../../../../../lib/infrastructure/repositories/complementary-certification-repository.js';
-import { CertificationCandidate } from '../../../../../lib/domain/models/CertificationCandidate.js';
+import { CertificationCandidate } from '../../../../../lib/domain/models/index.js';
 import { CertificationCandidatesError } from '../../../../../lib/domain/errors.js';
 import fs from 'fs';
 import _ from 'lodash';
