+++ api/tests/tooling/domain-builder/factory/build-country.js
@@ -1,5 +1,5 @@
 import { Country } from '../../../../lib/domain/read-models/Country.js';
-import { normalizeAndSortChars } from '../../../../lib/infrastructure/utils/string-utils.js';
+import { normalizeAndSortChars } from '../../../../src/shared/infrastructure/utils/string-utils.js';
 
 const buildCountry = function ({ code = '99345', name = 'TOGO', matcher = normalizeAndSortChars(name) } = {}) {
   return new Country({
